Revert default UUID to v6
Production version of uProtocol (1.3.7) used UUIDv6 and changing to v8
would result in a non-backwards compatible change as CloudEvents assume v6
because of the Validators that are used in devices.
In order to change versions of UUIDs, we'll need to update the message
type version (ex. pub.v1 to pub.v2), something to be considered at a
later date.

#15