Change up-java versions
the current up-java version of 1.5.x was out of sync with the up-core-api/up-spec version numbers so there was confusion on if the two are related meaning is 1.5.7 of up-java the same as 1.5.7 of up-core-api and they are not. In order to avoid confusion, we are renaming the up-java to 0.1.9 and we will add the up-core-api-1.5.7 tag as well so we know what version this support of up-core-api.