Remove CloudEvents Code (#122)
Cloudevents were used in older versions of the protocol but they have since been replaced with our own message format called UMessage. The cloudevent code was copied over to up-java so that we could support backwards compatibility however given that UUri and UUID changes, it was no longer possible to simply convert from 1.3.6 cloudevent to 1.5.x UMessage. After further study of deployed solutions, it was found that we can support the old CloudEvents and 1.3..x uEs (and deployments) using the older SDK so that this library does not require supporting CloudEvents at all. This commit removes all the CloudEvent code so that we no longer have to maintain it.

#121