Fix return types of async APIs
The operations of the L1, L2 and L3 APIs have all been returning CompletableFuture<UStatus>
and had always been succeeded with a UStatus. Client code then had to check if the UCode was
OK or not. This has been changed so that the operations now return CompletableFuture<Void>
and complete normally if the operation was successful. If the operation failed, the future
completes exceptionally with a UStatusException containing the UStatus with the error code.

This is much more aligned with the way async operations are usually handled in Java
and makes client code much cleaner.